Move album related code to own file
