+++ cmd/lastfm.go
@@ -5,10 +5,8 @@ import (
 	"errors"
 	"fmt"
 	"io"
-	"log"
 	"net/http"
 	"os"
-	"strconv"
 )
 
 type LastFMImage struct {
@@ -24,22 +22,6 @@ type LastFMUser struct {
 	Total      string `json:"total"`
 }
 
-type LastFMAlbum struct {
-	Artist struct {
-		URL        string `json:"url"`
-		ArtistName string `json:"name"`
-		Mbid       string `json:"mbid"`
-	} `json:"artist"`
-	Image     []LastFMImage `json:"image"`
-	Mbid      string        `json:"mbid"`
-	URL       string        `json:"url"`
-	Playcount string        `json:"playcount"`
-	Attr      struct {
-		Rank string `json:"rank"`
-	} `json:"@attr"`
-	AlbumName string `json:"name"`
-}
-
 type LastFMArtist struct {
 	Image     []LastFMImage `json:"image"`
 	Mbid      string        `json:"mbid"`
@@ -49,30 +31,6 @@ type LastFMArtist struct {
 		Rank string `json:"rank"`
 	} `json:"@attr"`
 	Name string `json:"name"`
-}
-
-type LastFMTopAlbums struct {
-	TopAlbums struct {
-		Album []LastFMAlbum `json:"album"`
-		Attr  LastFMUser    `json:"@attr"`
-	} `json:"topalbums"`
-}
-
-func (a *LastFMTopAlbums) Append(l LastFMResponse) {
-	if albums, ok := l.(*LastFMTopAlbums); ok {
-		a.TopAlbums.Album = append(a.TopAlbums.Album, albums.TopAlbums.Album...)
-		return
-	}
-	log.Println("Error: LastFMResponse is not a LastFMTopAlbums")
-}
-
-func (a *LastFMTopAlbums) GetTotalPages() int {
-	totalPages, _ := strconv.Atoi(a.TopAlbums.Attr.TotalPages)
-	return totalPages
-}
-
-func (a *LastFMTopAlbums) GetTotalFetched() int {
-	return len(a.TopAlbums.Album)
 }
 
 type LastFMTopArtists struct {
@@ -162,31 +120,4 @@ func getLastFmResponse[T LastFMResponse](collageType CollageType, username strin
 		page++
 	}
 	return &result, nil // No more pages to fetch
-}
-
-func getAlbums(collageType CollageType, username string, period Period, count int, imageSize string) ([]*Album, error) {
-
-	result, err := getLastFmResponse[*LastFMTopAlbums](collageType, username, period, count, imageSize)
-	if err != nil {
-		return nil, err
-	}
-	r := *result
-
-	albums := make([]*Album, len(r.TopAlbums.Album))
-	for i, album := range r.TopAlbums.Album {
-		newAlbum := &Album{
-			Name:      album.AlbumName,
-			Artist:    album.Artist.ArtistName,
-			Playcount: album.Playcount,
-		}
-
-		for _, image := range album.Image {
-			if image.Size == imageSize {
-				newAlbum.ImageUrl = image.Link
-			}
-		}
-
-		albums[i] = newAlbum
-	}
-	return albums, nil
 }
