--- cmd/models.go
@@ -1,57 +1,0 @@
-package main
-
-import "image"
-
-type Album struct {
-	Name      string
-	Artist    string
-	Playcount string
-	ImageUrl  string
-	Image     image.Image
-}
-
-func (a *Album) GetImageUrl() string {
-	return a.ImageUrl
-}
-
-func (a *Album) SetImage(img *image.Image) {
-	a.Image = *img
-}
-
-func (a *Album) GetImage() *image.Image {
-	return &a.Image
-}
-
-func (a *Album) GetParameters() map[string]string {
-	return map[string]string{
-		"artist":    a.Artist,
-		"album":     a.Name,
-		"playcount": a.Playcount,
-	}
-}
-
-type Artist struct {
-	Name      string
-	Playcount string
-	Image     image.Image
-	ImageUrl  string
-}
-
-func (a *Artist) GetImageUrl() string {
-	return a.ImageUrl
-}
-
-func (a *Artist) SetImage(img *image.Image) {
-	a.Image = *img
-}
-
-func (a *Artist) GetImage() *image.Image {
-	return &a.Image
-}
-
-func (a *Artist) GetParameters() map[string]string {
-	return map[string]string{
-		"artist":    a.Name,
-		"playcount": a.Playcount,
-	}
-}
